Shaping up code
