+++ mvg/features/rms.py
@@ -1,14 +1,7 @@
-import pickle
-import json
-import numpy as np
+"""Analysis Class for RMS Feature"""
 import pandas as pd
 import matplotlib.pyplot as plt
-from matplotlib import cm
-import matplotlib.patches as mpatches
 from tabulate import tabulate
-from mvg import plotting
-
-
 from mvg.features.analysis import Analysis
